Support jRuby gem installation
Configure the jRuby compatible gem version to build using the
ext/Rakefile default task. Extract the same logic for downloading the
extension and gem to a separate file so it can be easily reused.

Change the host OS detection from `Gem::Platform.local.os` to
`RbConfig::CONFIG["host_os"]` as `Gem::Platform.local.os` returns "java"
on jRuby, which isn't very helpful in detecting the system OS.
