+++ .rubocop.yml
@@ -54,6 +54,10 @@ Style/MethodMissing:
 Style/TrailingUnderscoreVariable:
   Enabled: false
 
+Style/FileName:
+  Exclude:
+    - "ext/Rakefile"
+
 Metrics/BlockLength:
   Exclude:
     - "Rakefile"
