+++ ext/Rakefile
@@ -1,0 +1,27 @@
+require File.expand_path("../../lib/appsignal/version.rb", __FILE__)
+require File.expand_path("../base.rb", __FILE__)
+
+task :default do
+  begin
+    logger.info "Installing appsignal agent #{Appsignal::VERSION} for Ruby #{RUBY_VERSION} on #{RUBY_PLATFORM}"
+    write_agent_architecture
+    next unless check_architecture
+    arch_config = AGENT_CONFIG["triples"][ARCH]
+
+    unless File.exist?(ext_path("appsignal-agent")) &&
+        (
+          File.exist?(ext_path("libappsignal.dylib")) ||
+          File.exist?(ext_path("libappsignal.so"))
+        ) &&
+        File.exist?(ext_path("appsignal.h"))
+      archive = download_archive(arch_config, "dynamic")
+      next unless verify_archive(archive, arch_config, "dynamic")
+      unarchive(archive)
+    end
+  rescue => ex
+    installation_failed "Exception while installing: #{ex}"
+    ex.backtrace.each do |line|
+      logger.error line
+    end
+  end
+end
