+++ lib/appsignal.rb
@@ -778,6 +778,7 @@ module Appsignal
   end
 end
 
+require "appsignal/system"
 require "appsignal/utils"
 require "appsignal/extension"
 require "appsignal/auth_check"
@@ -796,4 +797,3 @@ require "appsignal/rack/generic_instrumentation"
 require "appsignal/rack/js_exception_catcher"
 require "appsignal/js_exception_transaction"
 require "appsignal/transmitter"
-require "appsignal/system"
